% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_report.r
\name{build_report}
\alias{build_report}
\title{Build a report with basic spatio-temporal movement computations}
\usage{
build_report(df, path = ".", stats = c("rolling", "diurnal"),
  construct = c("klocoh"), proj4, seas = NULL, wavelet = T)
}
\arguments{
\item{df}{a dataframe containing columns "x", "y", and "date"}

\item{path}{output file path}

\item{stats}{a character vector of stats to calculate, options include: "rolling",
"diurnal", "lunar", and "seasonal"}

\item{construct}{a character vector indicating which spacetime construction methods to use.
options include "klocoh" and "akde"}

\item{proj4}{a character string indicating the proj.4 definition of the
coordinate reference system defining the relocations}

\item{seas}{a named numeric vector including the start date of each season
indicated by the Julian day (see \code{lubridate::yday} for easy conversion).
Required if \code{type == "seasonal"}}

\item{wavelet}{a logical indicating whether or not to produce wavelet visualization}
}
\description{
Build a report with basic spatio-temporal movement computations
}
\examples{
\donttest{
build_report(AG195, proj4 = "+proj=utm +zone=33 +south +datum=WGS84 +units=m +no_defs")
}
}
