% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructions.r
\name{construct}
\alias{construct}
\title{Home Range Construction by Two Methods}
\usage{
construct(df, type = c("klocoh", "akde"), proj4)
}
\arguments{
\item{df}{a dataframe containing columns x, y, date representing relocations in space and time.}

\item{type}{a character string indicating the type/s of constructions to build, "klocoh" or "akde"}

\item{proj4}{a character string indicating the proj.4 definition of the
coordinate reference system defining the relocations}
}
\description{
Home Range Construction by Two Methods
}
\examples{
\donttest{
klocoh <- construct(AG195, type = "klocoh", proj = "+proj=utm +zone=33 +south")
akde <- construct(AG195, type = "akde", proj = "+proj=utm +zone=33 +south")
}
}
