% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{create_telemetry}
\alias{create_telemetry}
\title{Build telemetry objects using projected data}
\usage{
create_telemetry(df, proj4)
}
\arguments{
\item{df}{a dataframe containing columns x, y, date representing relocations in space and time.}

\item{proj4}{a character string indicating the proj.4 definition of the
coordinate reference system defining the relocations}
}
\description{
A helper function to create telemetry objects required by \code{ctmm}
}
