% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df_check}
\alias{df_check}
\title{Check dataframes conform to stmove style}
\usage{
df_check(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
An (internal) helper function to check dataframes and provide helpful errors
}
\keyword{internal}
