% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{dist_wavelet}
\alias{dist_wavelet}
\title{Wavelet analyses}
\usage{
dist_wavelet(x, plot = T, na.approx = T)
}
\arguments{
\item{x}{a dataframe with columns: x, y, date, and id (optional)}

\item{plot}{a logical indicating whether or not to return a histogram of the distribution}

\item{na.approx}{a logical indicating whether NA values should be interpolated using to \code{zoo::na.approx}}
}
\description{
Wavelet analyses
}
\examples{
\donttest{
ss_dist(AG195)
ta_dist(AG195)
}


# Build morelet wavelet diagram based on step sizes over the time series
}
