% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{interval_stats}
\alias{interval_stats}
\title{Interval Summary Statistics}
\usage{
interval_stats(df, type = "diurnal", seas = NULL)
}
\arguments{
\item{df}{a dataframe containing columns "x", "y", and "date"}

\item{type}{a character string indicating the intervals to calculate, one of
"diurnal", "lunar", or "seasonal".}

\item{seas}{a character vector including the start date of each season interval in the format
\%Y-\%m-\%d, e.g. "2015-01-01".  Required if \code{type == "seasonal"}.}
}
\description{
Interval Summary Statistics
}
\details{
Note: if your trajectory includes dates before the first date in your \code{seas} vector, they
will automatically be considered a separate season.
}
\examples{
# diurnal is default
interval_stats(AG195)
interval_stats(AG195, type = "lunar")
# for seasonal, include y-m-d formatted `seas` vector
interval_stats(AG195, type = "seasonal", seas = c("2010-03-01", "2010-06-01", "2010-9-01"))
}
\seealso{
\link[lunar]{lunar.phase}
}
