% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{kalman}
\alias{kalman}
\title{Apply Kalman smoothing to fill gaps in data}
\usage{
kalman(df)
}
\arguments{
\item{df}{a dataframe containing columns x, y, date representing relocations in space and time.}
}
\description{
Apply Kalman smoothing to fill gaps in data
}
