% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{plot_timeline}
\alias{plot_timeline}
\title{Population Timeline Plot}
\usage{
plot_timeline(df)
}
\arguments{
\item{df}{a dataframe containing columns "x", "y", "date", and "id"}
}
\description{
Population Timeline Plot
}
