% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_report_addin.r
\name{report_addin}
\alias{report_addin}
\title{Build a Report (RStudio Add-in)}
\usage{
report_addin()
}
\description{
\code{report_addin()} opens an \href{https://shiny.rstudio.com/articles/gadgets.html}{RStudio gadget} and
\href{http://rstudio.github.io/rstudioaddins/}{addin} that allows you to
choose which analyses to run and build a report.
Appears as "Build report" in the RStudio Addins menu.
}
