% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{rolling_stats}
\alias{rolling_stats}
\title{Rolling Summary Statistics}
\usage{
rolling_stats(df)
}
\arguments{
\item{df}{a dataframe containing columns "x", "y", and "date}
}
\description{
Rolling Summary Statistics
}
\details{
some detailled discussion of how auto and cross correlation are calculated and
na handling
}
