% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.r
\name{ss_dist}
\alias{ss_dist}
\title{Basic Distributions}
\usage{
ss_dist(x)
}
\arguments{
\item{x}{a dataframe with columns: x, y, date, and id (optional)}
}
\value{
an ltraj object and a plot of the turning angle or step length distribution
}
\description{
Basic Distributions
}
\examples{
\donttest{
ss_dist(AG195)
ta_dist(AG195)
}

}
