% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_report.r
\name{build_report}
\alias{build_report}
\title{Build a report with basic spatio-temporal movement computations}
\usage{
build_report(file, df)
}
\arguments{
\item{file}{a file name and path.}

\item{df}{a dataframe containing columns x, y, date representing relocations in space and time.}
}
\description{
Build a report with basic spatio-temporal movement computations
}
