% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_report.r
\name{build_report}
\alias{build_report}
\title{Build a report with basic spatio-temporal movement computations}
\usage{
build_report(df, path, proj4, stats = c("rolling", "diurnal"),
  construct = c("klocoh"), seas = NULL, wavelet = NULL)
}
\arguments{
\item{df}{a dataframe containing columns "x", "y", and "date"}

\item{path}{absolute file path for output .pdf}

\item{proj4}{a character string indicating the proj.4 definition of the
coordinate reference system defining the relocations}

\item{stats}{a character vector of stats to calculate, options include: "rolling",
"diurnal", "lunar", and "seasonal"}

\item{construct}{a character vector indicating which spacetime construction methods to use.
options include "klocoh" and "akde"}

\item{seas}{a character vector including the start date of each season interval in the format
\%Y-\%m-\%d, e.g. "2015-01-01".  Required if \code{type == "seasonal"}.}

\item{wavelet}{a character vector indicating which variables to calculate wavelet analysis on,
options in "dist", "rel.angle", "acf_dist", "acf_ang", and "ccf".}
}
\description{
Build a report with basic spatio-temporal movement computations
}
\examples{
\donttest{
build_report(AG195, proj4 = "+proj=utm +zone=33 +south +datum=WGS84 +units=m +no_defs")
}
}
