% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{interval_stats}
\alias{interval_stats}
\title{Interval Summary Statistics}
\usage{
interval_stats(df, type = "diurnal", seas = NULL)
}
\arguments{
\item{df}{a dataframe containing columns "x", "y", and "date"}

\item{type}{a character string indicating the intervals to calculate, one of
"diurnal", "lunar", or "seasonal".}

\item{seas}{a named numeric vector including the start date of each season
indicated by the Julian day (see \code{lubridate::yday} for easy conversion).
Required if \code{type == "seasonal"}}
}
\description{
Interval Summary Statistics
}
