% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{kalman}
\alias{kalman}
\title{Kalman Smoothing with Structural Modeling}
\usage{
kalman(df)
}
\arguments{
\item{df}{a dataframe containing columns x, y, date representing relocations in space and time.}
}
\value{
a dataframe with additional binary column \code{real} flagging those points whose positions were interpolated.
}
\description{
Apply Kalman smoothing to interpolate missing values in trajectory.
}
\details{
The replacement points are generated using a structural model fitted by maximum likelihood.
}
\seealso{
\link[imputeTS]{na.kalman}
}
