% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.r
\name{ss_dist}
\alias{ss_dist}
\alias{ta_dist}
\title{Basic Distributions}
\usage{
ss_dist(x, plot = T)

ta_dist(x, plot = T)
}
\arguments{
\item{x}{a dataframe with columns: x, y, date, and id (optional)}

\item{plot}{a logical indicating whether or not to return a histogram of the distribution}
}
\value{
a numeric vector and (optionally) a plot of the turning angle or step length distribution
}
\description{
The first step to any movement analysis is to extract the step length and turning angle
distributions of trajectory. These functions will calculate step length and relative
turning angles for a given trajectory and plot associated histograms.
}
\examples{
\donttest{
ss <- ss_dist(AG195)
ta <- ta_dist(AG195)
}

}
