% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{wavelet}
\alias{wavelet}
\title{Wavelet analyses}
\usage{
wavelet(df, stats = c("dist", "rel.angle", "acf_dist", "acf_ang", "ccf"),
  plot = T)
}
\arguments{
\item{df}{a dataframe with columns: x, y, date}

\item{stats}{a character vector indicating which variables to calculate wavelet analysis on,
options in "dist", "rel.angle", "acf_dist", "acf_ang", and "ccf". By default, analysis is run on all 5.}

\item{plot}{a logical indicating whether or not to return a histogram of the distribution}
}
\description{
Build morelet wavelet diagrams for step size, turning angle, auto and cross correlation
coefficients through time.
}
\examples{
\donttest{
wavelet(AG195)
wavelet(AG195, c("dist", "angle"))
}

}
